FastAPI/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
python-jobspy>=1.1.0
pandas>=2.0.0
httpx>=0.25.0
boto3>=1.34.0
pytest>=8.0.0
pytest-cov>=5.0.0
